divide ssh connect steps
